+++ lang/system.go
@@ -10,22 +10,23 @@ import (
 	"github.com/markkurossi/iql/types"
 )
 
+// System variables.
 const (
-	sysRealFmt = "REALFMT"
-	sysTermOut = "TERMOUT"
+	SysRealFmt = "REALFMT"
+	SysTermOut = "TERMOUT"
 )
 
 // InitSystemVariables initializes the global system variables for the
 // scope.
 func InitSystemVariables(scope *Scope) {
-	scope.Declare(sysRealFmt, types.String)
-	scope.Declare(sysTermOut, types.Bool)
-	scope.Set(sysTermOut, types.BoolValue(true))
+	scope.Declare(SysRealFmt, types.String)
+	scope.Declare(SysTermOut, types.Bool)
+	scope.Set(SysTermOut, types.BoolValue(true))
 }
 
 // Format gets the value formatting options from the scope.
 func Format(scope *Scope) *types.Format {
-	real := scope.Get(sysRealFmt)
+	real := scope.Get(SysRealFmt)
 	if real == nil {
 		return nil
 	}
